Prevent daemon99 crash
`shutil.move` gives an error when trying to overwrite an existing file.
Don't move data if old data wasn't removed by server

+++ daemon99.py
@@ -76,17 +76,13 @@ def do_mv_data(rpath):
 	for file in glob.glob(r'/tmp/*.csv'):
 		#print file
 		if os.path.isfile(clientlock):
-			shutil.move(file, rpath)
-
-	#for file in glob.glob(r'/tmp/*.err'):
-	#	print file
-	#	if os.path.isfile(clientlock):
-	#		shutil.move(file, rpath)
+			if not (os.path.isfile(rpath + "/" + os.path.split(file)[1])):
+			  shutil.move(file, rpath)
 
 	for file in glob.glob(r'/tmp/*.png'):
-		#print file
 		if os.path.isfile(clientlock):
-			shutil.move(file, rpath)
+			if not (os.path.isfile(rpath + "/" + os.path.split(file)[1])):
+				shutil.move(file, rpath)
 
 	unlock(clientlock)
